% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_models.R
\name{add_custom_lv_model}
\alias{add_custom_lv_model}
\title{Add a custom latent variable model}
\usage{
add_custom_lv_model(model, lv_model_code, items = all_items(model))
}
\arguments{
\item{model}{An irt_model object}

\item{lv_model_code}{NM-TRAN abbreviated code as a character vector of lines}

\item{items}{A vector of item numbers}
}
\value{
A new irt_model object
}
\description{
The model will be used for the items with numbers listed in the items argument.
If a model was already added for a listed item the function will give an error.
Use reset_lv_models to restart adding models for the latent variable.
The model code should be in regular NM-TRAN format with the following restrictions:
\itemize{
\item The final line must define PSI. (i.e. PSI=some code)
\item ETAs and THETAs can be used and numbering of these must start from 1
}
}
