% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{add_estimation_options}
\alias{add_estimation_options}
\title{Add extra options to $ESTIMATION}
\usage{
add_estimation_options(model, options)
}
\arguments{
\item{model}{A model object}

\item{options}{Extra options to add to $EST as one string}
}
\value{
A new model object
}
\description{
Add extra options to $ESTIMATION
}
