% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_models.R
\name{add_lv_model}
\alias{add_lv_model}
\alias{lv_model_types}
\title{Add a pre-defined latent variable model}
\usage{
add_lv_model(model, lv_model_type = "constant", items = all_items(model))

lv_model_types 
#c("constant", "linear")
}
\arguments{
\item{model}{An irt_model object}

\item{lv_model_type}{The name of the type}

\item{items}{A vector of item numbers}
}
\value{
A new irt_model object
}
\description{
The specified latent variable model will be applied to all items listed. It can
be one of "constant" and "linear".
"constant" means PSI = THETA(1) + ETA(1)
"linear" means PSI = BASE + SLP * TIME where BASE = THETA(1) + ETA(1) and SLP = THETA(2) + ETA(2)
Each call to this function for a set of items will us a different set of ETAs and THETAs
The default is to use the same "constant" model for all items.
}
