% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{add_simulation}
\alias{add_simulation}
\title{Add simulation}
\usage{
add_simulation(model, nsim = 1)
}
\arguments{
\item{model}{A model object}

\item{nsim}{The number of simulations to add}
}
\value{
A new model object
}
\description{
Add a number of simulations to the model
}
