% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{all_items}
\alias{all_items}
\title{Get a vector of all item numbers from scale}
\usage{
all_items(scale)
}
\arguments{
\item{scale}{An irt_scale object}
}
\value{
A vector of all item numbers
}
\description{
Get a vector of all item numbers from scale
}
