% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_models.R
\name{all_items_have_lv_model}
\alias{all_items_have_lv_model}
\title{Check if all items have an attached lv model}
\usage{
all_items_have_lv_model(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A boolean
}
\description{
Check if all items have an attached lv model
}
