% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{all_parameter_names}
\alias{all_parameter_names}
\title{Get all unique non-ignored parameter names for a model}
\usage{
all_parameter_names(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A vector of parameter names
}
\description{
Get all unique non-ignored parameter names for a model
}
