% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{apply_consolidation}
\alias{apply_consolidation}
\title{Will remove the consolidated levels from the scale of a model}
\usage{
apply_consolidation(model)
}
\arguments{
\item{model}{irt_model object}
}
\value{
An irt_model object
}
\description{
Will remove the consolidated levels from the scale of a model
}
\keyword{internal}
