% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model}
\alias{bi_model}
\title{Create a bounded integer model object}
\usage{
bi_model(scale_or_levels)
}
\arguments{
\item{scale_or_levels}{A scale object or a total number of levels}
}
\value{
A model object
}
\description{
\code{bi_model} returns a newly created model object
}
