% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{binary_data_model_code}
\alias{binary_data_model_code}
\title{Generate NONMEM code for binary data items}
\usage{
binary_data_model_code()
}
\value{
A code generator
}
\description{
Generate NONMEM code for binary data items
}
