% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{binary_items}
\alias{binary_items}
\title{Get a list of all items of binary type from a scale}
\usage{
binary_items(scale)
}
\arguments{
\item{scale}{An irt_scale object}
}
\value{
A vector of item numbers for all binary items
}
\description{
Get a list of all items of binary type from a scale
}
