% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_irt_link.R
\name{calculate_cv_irt_information}
\alias{calculate_cv_irt_information}
\title{Calculate latent variable information for an IRT-informed CV model}
\usage{
calculate_cv_irt_information(res, psi_range = c(-4, 4))
}
\arguments{
\item{res}{Result object from an IRT link-analysis}

\item{psi_range}{Latent variable range}
}
\value{
A tibble
}
\description{
Calculate latent variable information for an IRT-informed CV model
}
