% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_irt_link.R
\name{calculate_cv_irt_link}
\alias{calculate_cv_irt_link}
\alias{plot_cv_irt_link}
\title{Determine IRT-based mean and SD link for CV model}
\usage{
calculate_cv_irt_link(
  model,
  psi_range = NULL,
  score_range = c(-Inf, Inf),
  range_tol = 0.3,
  approx_tol_mean = 0.1,
  approx_tol_sd = 0.01
)

plot_cv_irt_link(res)
}
\arguments{
\item{model}{}

\item{psi_range}{}

\item{score_range}{}

\item{range_tol}{}

\item{approx_tol_mean}{}

\item{approx_tol_sd}{}
}
\value{

}
\description{
Determine IRT-based mean and SD link for CV model
}
