% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_irt_link.R
\name{calculate_e_score_vs_psi}
\alias{calculate_e_score_vs_psi}
\alias{plot_e_score_vs_psi}
\alias{calculate_sd_score_vs_psi}
\alias{plot_sd_score_vs_psi}
\title{Calculate mean and SD as a function of PSI}
\usage{
calculate_e_score_vs_psi(model, psi_range = c(-4, 4))

plot_e_score_vs_psi(model, ...)

calculate_sd_score_vs_psi(model, psi_range = c(-4, 4))

plot_sd_score_vs_psi(model, ...)
}
\arguments{
\item{model}{An irt_model object}

\item{psi_range}{A vector of lenght 2 specifying the lower and upper boundary for the latent variable}
}
\value{
A data.frame or a plot
}
\description{
Calculate mean and SD as a function of PSI
}
