% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccs.R
\name{calculate_iccs}
\alias{calculate_iccs}
\alias{plot_iccs}
\title{Calculate and plot item characteristic curves}
\usage{
calculate_iccs(model, psi_range = c(-4, 4))

plot_iccs(model, ...)
}
\arguments{
\item{model}{An irt_model object}

\item{psi_range}{A vector of lenght 2 specifying the lower and upper boundary for the latent variable}
}
\value{
A data.frame or a plot
}
\description{
These functions calculate and plot the item characteristic curves for the model specified. The results are returned either 
as a data.frame (\code{calculate_iccs}) or visualized in a plot (\code{plot_iccs}).
}
