% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{calculate_item_information_curves}
\alias{calculate_item_information_curves}
\alias{plot_item_information_curves}
\title{Calculate item information curves}
\usage{
calculate_item_information_curves(model, psi_range = c(-4, 4))

plot_item_information_curves(model, psi_range = c(-4, 4))
}
\arguments{
\item{model}{An irt_model object}

\item{psi_range}{A vector of lenght 2 specifying the lower and upper boundary for the latent variable}
}
\value{
A data.frame or a plot
}
\description{
These functions calculate the item information curves for the model specified. The results are returned either 
as a data.frame (\code{calculate_item_information_curves}) or visualized in a plot (\code{plot_item_information_curves}).
}
