% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{calculate_population_item_information}
\alias{calculate_population_item_information}
\alias{plot_population_item_information}
\title{Calculate the item information in a population}
\usage{
calculate_population_item_information(model, mean = 0, var = 1)

plot_population_item_information(model, mean = 0, var = 1)
}
\arguments{
\item{model}{An irt_model object}

\item{mean}{mean of the population}

\item{var}{variance of the population}
}
\value{
A data.frame or a plot
}
\description{
These functions calculate the population item information for the model specified. The results are returned either 
as a data.frame (\code{calculate_population_item_information}) or visualized in a plot (\code{plot_population_item_information}).
}
