% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{calculate_wreslike}
\alias{calculate_wreslike}
\title{Calculate the WRESLIKE residual}
\usage{
calculate_wreslike(ptab)
}
\arguments{
\item{ptab}{Dataframe with the probability table from modelfit output}
}
\value{
The input table with WRESLIKE added
}
\description{
Calculate the WRESLIKE residual
}
