% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{consolidate_data}
\alias{consolidate_data}
\title{Apply the consolidations in model to the data by moving
consolidated levels into the nearest lower level.}
\usage{
consolidate_data(df, model)
}
\arguments{
\item{df}{A data.frame with ITEM and DV columns}

\item{model}{An irt_model object}
}
\value{
An updated data.frame
}
\description{
Apply the consolidations in model to the data by moving
consolidated levels into the nearest lower level.
}
