% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{consolidate_levels}
\alias{consolidate_levels}
\title{Consolidation of levels in a model}
\usage{
consolidate_levels(model, item_numbers, levels)
}
\arguments{
\item{model}{An irt_model object}

\item{item_numbers}{A vector of item numbers to consolidate}

\item{levels}{A vector of levels to consolidate, i.e. merging into the nearest lower level.}
}
\value{
A new irt_model object
}
\description{
The consolidated levels will be treated as being equal to the nearest lower level.
}
