% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{consolidate_levels_below}
\alias{consolidate_levels_below}
\title{Automatically consolidate all items given a certain minimum count}
\usage{
consolidate_levels_below(model, count)
}
\arguments{
\item{model}{An irt_model object}

\item{count}{The maximum count for consolidation}
}
\value{
A new irt_model object
}
\description{
If a the occurences of a level for an item is not above count it will
be consolidated into the nearest level. Consolidation can only be done
from the higher levels and down using this function. Levels that are
present in the scale but missing in the dataset also gets consolidated.
}
