% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{consolidate_levels_in_model}
\alias{consolidate_levels_in_model}
\title{Consolidation of levels in a model without changing the scale}
\usage{
consolidate_levels_in_model(model, item_numbers, levels)
}
\arguments{
\item{model}{An irt_model object}

\item{item_numbers}{A vector of item numbers to do the same consolidation for}

\item{levels}{A vector of levels to consolidate, i.e. merging into the lowest level of this vector.}
}
\value{
A new irt_model object
}
\description{
When consolidating levels in the higher end of the levels it is possible to 
do it without altering the scale or the model. It is done by setting the inital
estimate of the DIF thetas to a high fixed number. This will make the probability
of the level to become zero. The benefit of doing this is that the same model can
be used for different datasets and different consolidations can be done simply by
changing the initial estimates.
}
