% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{consolidated}
\alias{consolidated}
\title{Check if a level for an item was consolidated}
\usage{
consolidated(model, item, level)
}
\arguments{
\item{model}{An irt_model object}

\item{item}{An irt_item object}

\item{level}{A level}
}
\value{
TRUE if level was consolidated for item and FALSE otherwise
}
\description{
Check if a level for an item was consolidated
}
