% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piraid_mirt_interop.R
\name{convert_to_long_nmdata}
\alias{convert_to_long_nmdata}
\title{Convert data from wide to long format}
\usage{
convert_to_long_nmdata(data)
}
\arguments{
\item{data}{A data.frame or matrix where each column corresponds to an item and each row
to a subject}
}
\value{
A data.frame with one row per subject and item
}
\description{
This functionc converts a data.frame in wide format to the long format.
}
