% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{correlation_plot}
\alias{correlation_plot}
\title{Item response correlation plot}
\usage{
correlation_plot(df)
}
\arguments{
\item{df}{A data.frame from the item_parameters_tab of a model run. ID, ITEM, TIME and PWRES will be used.}
}
\value{
A plot object
}
\description{
Item response correlation plot
}
