% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piraid_mirt_interop.R
\name{create_mirt_pseudo_data}
\alias{create_mirt_pseudo_data}
\title{Create dataset with all necessary responses}
\usage{
create_mirt_pseudo_data(model)
}
\arguments{
\item{model}{An irt_model}
}
\value{
data.frame
}
\description{
Create dataset with all necessary responses
}
