% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{create_scale_from_df}
\alias{create_scale_from_df}
\alias{create_scale_from_csv}
\title{Infer a scale from a dataset}
\usage{
create_scale_from_df(df, item = "ITEM", dv = "DV", name = NULL,
  type = NULL)

create_scale_from_csv(file, item = "ITEM", dv = "DV", name = NULL,
  type = NULL)
}
\arguments{
\item{df}{A data.frame}

\item{item}{Name of the item columns. Default is \code{ITEM}}

\item{dv}{Name of the dv column. Default is \code{DV}}

\item{name}{Name of an optional name column. Each item can be given a name (mainly for plotting) through this column. Only the first row of a new item will be used for the name.}

\item{type}{Name of an optional type column. Each item can be given a type (either \code{item_type$ordered_categorical} or \code{item_type$binary}) through this column. If the type option isn't used the type is 
inferred from the number of distinct DV values}

\item{file}{The path to a csv file}
}
\value{
An irt_scale object
}
\description{
These functions allow to infer a scale definition from a dataset (provided as data.frame or CSV file) if it follows a general data specification template.
}
\details{
In order for the automatic inference to work, the dataset needs to contain the columns ITEM and DV. The function will then use the observed levels for each 
item to determine the item type (binary or ordered categorical) while taking MDV and EVID columns into account (records with MDV or EVID not equal to 0 are
ignored) . The automatic recognition can be over-written by providing a type column which can specify the item type using the keywords 'ordcat' and 'binary'.
These keywords are available through the item_type list, i.e. \code{item_type$ordered_categorical} and \code{item_type$binary}.
Additionally, a label for each item can be provided through the column name.
}
