% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{create_subscale}
\alias{create_subscale}
\title{Create subscale}
\usage{
create_subscale(scale, categories)
}
\arguments{
\item{scale}{An irt_scale object}

\item{categories}{A vector of category names to include in the new scale}
}
\value{
A new irt_scale object
}
\description{
\code{create_subscale} Create a new scale with items from selected categories only
}
