% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_model.R
\name{cv_model}
\alias{cv_model}
\title{Create a continuous variable model object}
\usage{
cv_model(scale_or_min, max = NULL)
}
\arguments{
\item{scale_or_min}{A scale object or a minimum total score}

\item{max}{A maximum total score}
}
\value{
A model object
}
\description{
\code{cv_model} returns a newly created model object
}
