% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{data_check}
\alias{data_check}
\title{Check dataset for missing items or levels}
\usage{
data_check(model_or_data, scale = NULL)
}
\arguments{
\item{model_or_data}{Either a data.frame or a model object from which to take the dataset}

\item{scale}{If model_or_data was a data.frame a scale must be supplied otherwise the scale from the model object will be taken}
}
\description{
The check will check for the following:
1. Items  present in the dataset but not in the scale
2. Items present in the scale but not in the dataset
3. Levels present in the dataset but not in the scale
4. Levels present in the scale but not in the dataset
The results of the check will be printed to the console
}
