% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{data_models_code}
\alias{data_models_code}
\title{Generate NONMEM code for all different item model types}
\usage{
data_models_code(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A code generator object
}
\description{
Generate NONMEM code for all different item model types
}
