% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{diagnose_icc_fit}
\alias{diagnose_icc_fit}
\alias{icc_plots}
\title{Plot item characteristic curves diagnostic}
\usage{
diagnose_icc_fit(
  model,
  nmtab,
  psi_range = c(-4, 4),
  resample_psi = FALSE,
  samples = 10,
  items_per_page = NULL
)

icc_plots(
  model,
  nmtab,
  psi_range = c(-4, 4),
  resample_psi = FALSE,
  samples = 10,
  items_per_page = NULL
)
}
\arguments{
\item{model}{The model that was used to create the output table}

\item{nmtab}{A data.frame from the item_parameters_tab of a model run. PSI, ITEM, DV, DIS, DIFn are needed}

\item{psi_range}{The range of psi values to use for the plot}

\item{resample_psi}{Whether to use the resampling based diagnostic}

\item{samples}{The number of samples to use when resample_psi = T}

\item{items_per_page}{The number of items to display on one page (default NULL prints all items)}
}
\value{
A list of plots
}
\description{
MDV and EVID will be used to filter out non-observations before plotting
}
