% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\docType{data}
\name{estimate_item_parameters}
\alias{estimate_item_parameters}
\alias{data_use_strategies}
\title{Estimate item parameter}
\format{An object of class \code{character} of length 2.}
\usage{
estimate_item_parameters(model, data_use_strategy = "baseline")

data_use_strategies
}
\arguments{
\item{model}{The model}

\item{data_use_strategy}{The method of how to use the available longitudinal data}
}
\value{
A dataframe with the estimated item parameters
}
\description{
This function estimates the item parameters in a model using the \code{mirt} package. The dataset and the item models 
are taken from the provided model object.
}
\details{
The available strategies of how to use the longitudinal data include 'baseline' (the default) and 'visits-as-subjects'.

With the 'baseline' strategy only the baseline data is used for the estimation and all other data is ignored.

The 'visits-as-subjects' strategy uses all available data and treats each visit as a seperate subject during estimation.
}
\keyword{datasets}
