% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_lv_values}
\alias{estimate_lv_values}
\title{Estimate Latent Variable Values}
\usage{
estimate_lv_values(model,
  estimate_item_prms = !has_all_initial_estimates(model))
}
\arguments{
\item{model}{The model}

\item{estimate_item_prms}{Whether to re-estimate the item parameters from the data.}
}
\value{
A tibble with columns PSI (the estimated latent variable value), SE_PSI (the associated standard error), and all columns 
from the data (except ITEM and DV)
}
\description{
The function uses the data associated with the model to estimate a latent variable value for each time point.
}
