% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{estimation_task}
\alias{estimation_task}
\title{Generate NONMEM code for estimation and table output}
\usage{
estimation_task(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A code generator object
}
\description{
Generate NONMEM code for estimation and table output
}
