% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{fix_all_item_parameters}
\alias{fix_all_item_parameters}
\title{Fix all item parameters}
\usage{
fix_all_item_parameters(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
An updated irt_model object
}
\description{
The function fixes all item parameters in the model
}
