% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{fix_item_parameters}
\alias{fix_item_parameters}
\title{Fix the same parameters for some items}
\usage{
fix_item_parameters(model, items, parameter_names)
}
\arguments{
\item{model}{An irt_model object}

\item{items}{A vector of item numbers}

\item{parameter_names}{A vector of names of parameters}
}
\value{
A new irt_model object
}
\description{
Fix the same parameters for some items
}
