% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{format_integers}
\alias{format_integers}
\title{Format a vector containing integers to a readable string}
\usage{
format_integers(x)
}
\arguments{
\item{x}{A vector containing integers}
}
\value{
A readable string
}
\description{
The string will use a-b if there is a range of consecutive integers
between a and b inclusive. Commas will be used between such groups.
A lone integer will be written by it self.
}
\examples{
nmIRT:::format_integers(c(1,2,3,5,7,8)) #will return "1-3, 5, 7-8"

}
