% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{get_ignored_items}
\alias{get_ignored_items}
\title{Get a vector of all ignored items}
\usage{
get_ignored_items(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
Vetor of ignored items
}
\description{
Get a vector of all ignored items
}
