% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{get_item}
\alias{get_item}
\title{Get an item from a scale}
\usage{
get_item(scale, number)
}
\arguments{
\item{scale}{The scale}

\item{number}{The item number to retrieve}
}
\value{
The item with corresponding number or NULL if not found
}
\description{
\code{get_item} returns the item with the specified number (id)
}
\keyword{internal}
