% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{get_item_index}
\alias{get_item_index}
\title{Gets the vector index of an item}
\usage{
get_item_index(scale, number)
}
\arguments{
\item{scale}{An irt_scale object}

\item{number}{The item number of an item}
}
\value{
The vector index of the scale$item vector
}
\description{
Gets the vector index of an item
}
\keyword{internal}
