% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_models.R
\name{get_item_lv_index}
\alias{get_item_lv_index}
\title{Get the index of the latent variable model to use for a certain item}
\usage{
get_item_lv_index(model, item)
}
\arguments{
\item{model}{An irt_model object}

\item{item}{An item number}
}
\value{
The index
}
\description{
The returned index is for the internal model$lv_models list, but
also for the number of the PSI model in the NONMEM model. A returned
value if 0 means to use the default.
}
