% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piraid_mirt_interop.R
\name{get_mirt_names}
\alias{get_mirt_names}
\title{Get mirt names for all items in model}
\usage{
get_mirt_names(model)
}
\arguments{
\item{model}{An irt_model}
}
\value{
Vector of names for all items in model
}
\description{
Get mirt names for all items in model
}
