% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piraid_mirt_interop.R
\name{get_mirt_types}
\alias{get_mirt_types}
\title{Get mirt types for all items in model}
\usage{
get_mirt_types(model)
}
\arguments{
\item{model}{An irt_model}
}
\value{
Vector of mirt itemtypes
}
\description{
Get mirt types for all items in model
}
