% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_models.R
\name{get_number_of_thetas_and_etas}
\alias{get_number_of_thetas_and_etas}
\title{Get number of THETAs and ETAs used in a vector of code lines}
\usage{
get_number_of_thetas_and_etas(lines)
}
\arguments{
\item{lines}{A character vector of code lines}
}
\value{
A list with numetas and numthetas
}
\description{
Get number of THETAs and ETAs used in a vector of code lines
}
