% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{graded_response_model}
\alias{graded_response_model}
\title{Graded response model for an unknown number of DIF columns}
\usage{
graded_response_model(data)
}
\arguments{
\item{data}{A dataset the columns DIS, PSI, DIFi and CAT (level) will be used}
}
\value{
Vector of response values
}
\description{
Graded response model for an unknown number of DIF columns
}
\keyword{internal}
