% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{has_all_initial_estimates}
\alias{has_all_initial_estimates}
\title{Test if a model has all required inital estimates}
\usage{
has_all_initial_estimates(model)
}
\arguments{
\item{model}{The model}
}
\value{
TRUE if an initial value is given for all item parameters in the model, otherwise FALSE
}
\description{
Test if a model has all required inital estimates
}
