% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{ignore_items}
\alias{ignore_items}
\title{Ignore all parameters for some items}
\usage{
ignore_items(model, items)
}
\arguments{
\item{model}{An irt_model object}

\item{items}{A vector of item numbers}
}
\value{
A new irt_model objet
}
\description{
Ignoring an item means to set its parameters to 0 FIX
and to let data records with this item be ignored.
}
