% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{initial_estimate}
\alias{initial_estimate}
\title{Get the initial estimate for a certain item and parameter}
\usage{
initial_estimate(model, item, parameter)
}
\arguments{
\item{model}{An irt_model object}

\item{item}{An irt_item object}

\item{parameter}{The name of a parameter}
}
\value{
The initial estimate of the parameter for the item
}
\description{
The initial estimate will be:
}
\details{
\enumerate{
\item The user defined value from set_item_parameters
\item 50 if the parameter was consolidated
\item Taken from the published model in the scale
\item A fall back value of 0.1
}
}
