% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{initial_estimates_from_data}
\alias{initial_estimates_from_data}
\title{Calculate initial estimates given the data}
\usage{
initial_estimates_from_data(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A new irt_model
}
\description{
Use the mirt package to calculate initial estimates for the item parameters
Can only handle models where there are no missing levels except for at the high end of the range.
}
