% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{initial_estimates_item_parameters}
\alias{initial_estimates_item_parameters}
\title{Set initial estimates for some parameters}
\usage{
initial_estimates_item_parameters(model, items, parameters, inits)
}
\arguments{
\item{model}{An irt_model object}

\item{items}{A vector of item numbers}

\item{parameters}{A vector of parameter names}

\item{inits}{A vector of initial estimates. Must be equal in size to parameters or have size one.}
}
\value{
A new irt_model object
}
\description{
Each item will for the selected parameters get a new initial estimate.
If there is only one initial estimate provided all parameters will get this value.
If more than one initial estimate is provided they will be distributed to the parameters
in order. Note that the number of parameters and the number of initial estimates provided
must then be equal.
}
