% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{initial_item_thetas}
\alias{initial_item_thetas}
\title{Generate NONMEM code for initial estimates of thetas}
\usage{
initial_item_thetas(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A code generator object
}
\description{
Will use the initial estimates of the scale
}
