% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{insert_into_parameter_table}
\alias{insert_into_parameter_table}
\title{Insert into parameter table}
\usage{
insert_into_parameter_table(model, new_df, column)
}
\arguments{
\item{model}{An irt_model object}

\item{new_df}{Data to be inserted}

\item{column}{Name of the column to be inserted (fix or init)}
}
\value{
A new irt_model object
}
\description{
A helper function to insert new data into the parameter. The columns are:
item, parameter, fix and init
}
