% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{irt_data_and_input_code}
\alias{irt_data_and_input_code}
\title{Create the $DATA and $INPUT to NONMEM model code}
\usage{
irt_data_and_input_code(model, rewind = FALSE)
}
\arguments{
\item{model}{A irt_model object}

\item{rewind}{If the dataset should be rewound or not}
}
\value{
A code generator object
}
\description{
Create the $DATA and $INPUT to NONMEM model code
}
