% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{irt_item}
\alias{irt_item}
\title{Constructor for the irt_item class}
\usage{
irt_item(number, name, levels, type, categories = NULL, inits = NULL)
}
\arguments{
\item{number}{Item number}

\item{name}{A descriptive name of the item}

\item{levels}{A vector of levels for this item}

\item{categories}{An optional character vector of names of categories}

\item{inits}{Option vector of initial values for the item parameters}
}
\description{
Constructor for the irt_item class
}
\keyword{interal}
