% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{irt_item_assignment_fallthrough}
\alias{irt_item_assignment_fallthrough}
\title{Create fallthrough NONMEM code for unknown items}
\usage{
irt_item_assignment_fallthrough(cg, scale)
}
\arguments{
\item{cg}{A code generator object to build upon}

\item{scale}{An irt_scale object}
}
\value{
A code generator object
}
\description{
If the dataset contains items that the model wasn't created
for this code takes care of bailing out. It also contains
some dummy assignments that are only there to silence some
NM-TRAN warnings.
}
