% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{irt_model}
\alias{irt_model}
\title{Create an IRT model object}
\usage{
irt_model(scale, base_scale)
}
\arguments{
\item{scale}{The scale}

\item{base_scale}{A base scale on which the scale is based}
}
\value{
A model object
}
\description{
\code{irt_model} returns a newly created model object
}
