% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{is.base_model}
\alias{is.base_model}
\title{Check wheter x is a base_model object}
\usage{
is.base_model(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
True if x is an base_model object
}
\description{
Check wheter x is a base_model object
}
