% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{is.irt_model}
\alias{is.irt_model}
\title{Check wheter x is an irt_model object}
\usage{
is.irt_model(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
True if x is an irt_model object
}
\description{
Check wheter x is an irt_model object
}
