% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{is.irt_scale}
\alias{is.irt_scale}
\title{Check wheter x is an irt_scale object}
\usage{
is.irt_scale(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
True if x is an irt_scale object
}
\description{
Check wheter x is an irt_scale object
}
