% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{is_item_parameter_fixed}
\alias{is_item_parameter_fixed}
\title{Check if an item parameter is fixed}
\usage{
is_item_parameter_fixed(model, item, parameter)
}
\arguments{
\item{model}{An irt_model object}

\item{item}{An irt_item object}

\item{parameter}{The item parameter name}
}
\value{
A string containing "FIX" or an empty string
}
\description{
True should be returned if:
}
\details{
\itemize{
\item init = 0
\item The level corresponding to param was consolidated (in the upper end of the levels)
\item fix was set using the fix_item_parameters function
}
}
