% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{item_inits}
\alias{item_inits}
\title{Create initial estimates $THETAs for item}
\usage{
item_inits(model, item, start_theta)
}
\arguments{
\item{model}{An irt_model object}

\item{item}{An irt_item object}

\item{start_theta}{The number of the first theta for this item}
}
\value{
A vector of initial estimates definitions for NONMEM
}
\description{
Create initial estimates $THETAs for item
}
\keyword{internal}
