% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{item_labels}
\alias{item_labels}
\title{Get vector of labels for parameters of item}
\usage{
item_labels(item)
}
\arguments{
\item{item}{An irt_item object}
}
\value{
A label vector
}
\description{
Get vector of labels for parameters of item
}
\keyword{interal}
