% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{item_level_count}
\alias{item_level_count}
\title{Get the count of each item and level}
\usage{
item_level_count(model_or_data)
}
\arguments{
\item{model_or_data}{An irt_model object or a data.frame}
}
\value{
data.frame with ITEM, DV and count columns
}
\description{
Get the count of each item and level
}
