% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{item_name_list}
\alias{item_name_list}
\title{Create a vector of item labels with item numbers as names of entries}
\usage{
item_name_list(scale)
}
\arguments{
\item{scale}{An irt_scale object}
}
\value{
A vector with item numbers as row names
}
\description{
Create a vector of item labels with item numbers as names of entries
}
\keyword{internal}
