% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{item_parameter_names}
\alias{item_parameter_names}
\title{Get vector of parameter names of an item}
\usage{
item_parameter_names(model, item_number)
}
\arguments{
\item{model}{An irt_model object}

\item{item_number}{An item number}
}
\value{
A character vector of parameter names
}
\description{
Get vector of parameter names of an item
}
