% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{item_probability_sum}
\alias{item_probability_sum}
\title{Generate a NONMEM code string with a sum of probabilities}
\usage{
item_probability_sum(levels)
}
\arguments{
\item{levels}{A vector of levels}
}
\value{
A string with NONMEM code for
}
\description{
Generate a NONMEM code string with a sum of probabilities
}
