% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\docType{data}
\name{item_type}
\alias{item_type}
\title{A list enumerating all supported item types}
\format{
An object of class \code{list} of length 2.
}
\usage{
item_type
}
\description{
A list enumerating all supported item types
}
\examples{
item_type$binary

}
\keyword{datasets}
