% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{items_by_type}
\alias{items_by_type}
\title{Get a vector of all items of certain type}
\usage{
items_by_type(scale, types)
}
\arguments{
\item{scale}{An irt_scale object}

\item{types}{A vector of item types}
}
\description{
Get a vector of all items of certain type
}
