% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{items_in_categories}
\alias{items_in_categories}
\title{Get a list of all items in one or more categories}
\usage{
items_in_categories(scale, categories)
}
\arguments{
\item{scale}{An irt_scale object}

\item{categories}{A vector of names of categories}
}
\value{
A vector of item numbers
}
\description{
Get a list of all items in one or more categories
}
