% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generator.R
\name{join_with_max_length}
\alias{join_with_max_length}
\title{Join strings together into lines. Each line cannot be longer than max.}
\usage{
join_with_max_length(strs, sep = " ", max = 100)
}
\arguments{
\item{strs}{A vector of strings to join together.}

\item{sep}{Separator for joining the strings.}

\item{max}{Maximum number of characters for each line.}
}
\value{
A vector of strings. Each element represents a line of text.
}
\description{
Join strings together into lines. Each line cannot be longer than max.
}
