% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_models.R
\name{latent_variable_code}
\alias{latent_variable_code}
\title{Generate abbreviated code for the latent variable model}
\usage{
latent_variable_code(model, next_theta)
}
\arguments{
\item{model}{An irt_model object}

\item{next_theta}{The number of the next theta to be used}
}
\value{
A list of: a code generator object, next theta and next eta
}
\description{
Generate abbreviated code for the latent variable model
}
