% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{levels_as_string}
\alias{levels_as_string}
\title{Take a vector of levels and produce a description string usable for presentation or output to scale file}
\usage{
levels_as_string(levels)
}
\arguments{
\item{levels}{A vector of levels}
}
\value{
A string describing the levels
}
\description{
Take a vector of levels and produce a description string usable for presentation or output to scale file
}
\keyword{internal}
