% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{list_initial_estimates}
\alias{list_initial_estimates}
\title{Generate a data frame to give an overview of initial estimates for a model}
\usage{
list_initial_estimates(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A data frame with columns item, parameter, fix and init for all items and parameters
}
\description{
Generate a data frame to give an overview of initial estimates for a model
}
