% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{list_items}
\alias{list_items}
\alias{summary.irt_scale}
\title{List available items in a scale}
\usage{
list_items(scale)

\method{summary}{irt_scale}(object, ...)
}
\arguments{
\item{scale}{An irt_scale object}

\item{object}{An irt_scale object}

\item{...}{No additional arguments are supported}
}
\value{
A data.frame with scale information
}
\description{
Generate a table with the columns Item, Levels, Type, Categories and Name to give an
overview of what a scale object contains.
}
