% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{list_predefined_scales}
\alias{list_predefined_scales}
\title{Retrieve a list of all built in scales}
\usage{
list_predefined_scales()
}
\value{
A character vector with the names
}
\description{
Retrieve a list of all built in scales
}
