% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{load_scale}
\alias{load_scale}
\title{Load a scale from file}
\usage{
load_scale(path)
}
\arguments{
\item{path}{Path to the file to read}
}
\value{
A scale object
}
\description{
\code{load_scale} loads a scale file from disk and returns a scale object
}
