% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lv_vs_time_plot}
\alias{lv_vs_time_plot}
\title{Plot latent variable versus time}
\usage{
lv_vs_time_plot(
  lv_dataset,
  lv_col = "PSI",
  time_col = "TIME",
  se_col = "SE_PSI",
  grouping_col = NULL
)
}
\arguments{
\item{lv_dataset}{A dataframe}

\item{lv_col}{Name of the latent variable column}

\item{time_col}{Name of the time column}

\item{se_col}{Name of the standard error of PSI column}

\item{grouping_col}{Name of optional grouping column}
}
\value{
A ggplot plot object
}
\description{
Plot latent variable versus time
}
