% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piraid_mirt_interop.R
\name{make_mirt_model}
\alias{make_mirt_model}
\title{Convert piraid model to mirt model}
\usage{
make_mirt_model(model)
}
\arguments{
\item{model}{A piraid irt_model}
}
\value{
An mirt SingleGroupClass
}
\description{
This function converts a piraid irt_model to an mirt SingleGroupClass to allow the subsequent use of mirt functions
}
