% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{mdv_string}
\alias{mdv_string}
\title{Get an MDV string for $TABLE if dataset has MDV column}
\usage{
mdv_string(model)
}
\arguments{
\item{model}{irt_model object}
}
\value{
An empty string or " MDV "
}
\description{
Get an MDV string for $TABLE if dataset has MDV column
}
\keyword{internal}
