% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{model_complete}
\alias{model_complete}
\title{Check if all mandatory components has been added to a model}
\usage{
model_complete(model)
}
\arguments{
\item{model}{irt_model object}
}
\description{
Currently the scale and the dataset are mandatory.
Will give error if model is not complete
}
\keyword{internal}
