% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{model_type_constant}
\alias{model_type_constant}
\title{Get code symbol for type constant given scale and item}
\usage{
model_type_constant(scale, item)
}
\arguments{
\item{scale}{An irt_scale object}

\item{item}{An irt_item object}
}
\value{
The code symbol for item (i.e. BIN, OC1 etc)
}
\description{
Get code symbol for type constant given scale and item
}
