% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{number_of_levels}
\alias{number_of_levels}
\title{Calculate the total number of levels for a scale}
\usage{
number_of_levels(scale)
}
\arguments{
\item{scale}{An irt_scale object}
}
\value{
The total number of scale levels
}
\description{
Calculate the total number of levels for a scale
}
