% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{omegas}
\alias{omegas}
\title{Generate NONMEM code for the random effect}
\usage{
omegas(model, numomegas)
}
\arguments{
\item{model}{An irt_model object}

\item{numomegas}{The number of omegas needed}
}
\value{
A code generator object
}
\description{
Generate NONMEM code for the random effect
}
