% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{ordcat_levels}
\alias{ordcat_levels}
\title{Get a sorted array of sizes of ordered categorical items}
\usage{
ordcat_levels(scale)
}
\arguments{
\item{scale}{A scale object}
}
\description{
Will give a list of all unique number of items for all ordcat items
}
\keyword{internal}
