% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{ordered_categorical_data_model_code}
\alias{ordered_categorical_data_model_code}
\title{Generate NONMEM code for an ordered categorical item model}
\usage{
ordered_categorical_data_model_code(scale, levels)
}
\arguments{
\item{scale}{A scale object}

\item{levels}{A vector of the response levels}
}
\value{
A code generator object
}
\description{
Generate NONMEM code for an ordered categorical item model
}
