% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{ordered_categorical_simulation_code}
\alias{ordered_categorical_simulation_code}
\title{Generate NONMEM code for simulation of and ordered categorical item}
\usage{
ordered_categorical_simulation_code(scale, levels)
}
\arguments{
\item{scale}{A scale object}

\item{levels}{A vector of levels for the item}
}
\value{
A code generator object
}
\description{
Generate NONMEM code for simulation of and ordered categorical item
}
