% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piraid.R
\docType{package}
\name{piraid-package}
\alias{piraid}
\alias{piraid-package}
\title{pidaid: An aid for development and diagnosis of pharmacometric IRT models}
\description{
Automatic generation of different types of item response theory (IRT) models.
    Diagnostic tools for IRT modelling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UUPharmacometrics/piraid}
  \item Report bugs at \url{https://github.com/UUPharmacometrics/piraid/issues}
}

}
\author{
\strong{Maintainer}: Rikard Nordgren \email{rikard.nordgren@farmbio.uu.se} [copyright holder]

Authors:
\itemize{
  \item Sebastian Ueckert [copyright holder]
  \item Mats O. Karlsson [copyright holder]
}

Other contributors:
\itemize{
  \item Leticia Arrington [contributor]
}

}
