% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{predefined_scale}
\alias{predefined_scale}
\title{Retrieve a built in scale}
\usage{
predefined_scale(scale_name)
}
\arguments{
\item{scale_name}{The name of the scale}
}
\value{
A scale object or an error if the scale does not exist
}
\description{
\code{predefined_scale} returns a scale object from a built in scale.
Use \code{\link{list_predefined_scales}} to get a list of all available scales.
}
