% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_analysis.R
\name{prepare_lv_analysis}
\alias{prepare_lv_analysis}
\title{Prepare latent variable analysis}
\usage{
prepare_lv_analysis(lv_dataset, output_directory,
  model_filename = "lv_run1.mod", data_filename = "lv_data.csv")
}
\arguments{
\item{lv_dataset}{The dataset with the latent variable}

\item{output_directory}{The directory of where to store the NONMEM control stream and the dataset}

\item{model_filename}{The filename for the NONMEM control stream that will be created}

\item{data_filename}{The filename for the CSV data file that will be created}
}
\description{
The function accepts a dataset with a column containing latent variable estimates and uses it to prepare a direct latent variable analysis in NONMEM by
generating scaffold NONMEM model and dataset in the desired directory.
}
