% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piraid_mirt_interop.R
\name{prepare_mirt_type_vector}
\alias{prepare_mirt_type_vector}
\title{Generate mirt type vector for irt_model object}
\usage{
prepare_mirt_type_vector(model, wide_data)
}
\arguments{
\item{model}{irt_model object}

\item{wide_data}{data in wide format}
}
\value{
a vector with item model types for mirt
}
\description{
Generate mirt type vector for irt_model object
}
