% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{print.irt_model}
\alias{print.irt_model}
\title{Print a brief overview of a model}
\usage{
\method{print}{irt_model}(x, ...)
}
\arguments{
\item{x}{An irt_model object}

\item{...}{No additional arguments are supported}
}
\description{
Print a brief overview of a model
}
