% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{print.irt_scale}
\alias{print.irt_scale}
\title{Print a short overview of a scale}
\usage{
\method{print}{irt_scale}(x, ...)
}
\arguments{
\item{x}{An irt_scale object}

\item{...}{No additional arguments are supported}
}
\description{
Print a short overview of a scale
}
