% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{print_model_code}
\alias{print_model_code}
\title{Print NONMEM code}
\usage{
print_model_code(model)
}
\arguments{
\item{model}{A model object}
}
\description{
Output the NONMEM code of a model object to the console
}
