% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{print_scale_info}
\alias{print_scale_info}
\title{Print a short overview of a scale}
\usage{
print_scale_info(scale, header = TRUE)
}
\arguments{
\item{scale}{An irt_scale object}

\item{header}{Set to print the header (set to FALSE if called from print.irt_model)}
}
\description{
Print a short overview of a scale
}
