% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{published_init}
\alias{published_init}
\title{Get published initial estimate for a parameter of an item}
\usage{
published_init(scale, item_number, parameter_index)
}
\arguments{
\item{scale}{An irt_scale object}

\item{item_number}{The number of an item}

\item{parameter_index}{The index of a parameter}
}
\value{
The initial estimate or NULL if not available
}
\description{
Get published initial estimate for a parameter of an item
}
\keyword{internal}
