% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_import.R
\name{read_item_parameters_from_nmtab}
\alias{read_item_parameters_from_nmtab}
\title{Read item paramter values from NM table}
\usage{
read_item_parameters_from_nmtab(model, nmtab_file)
}
\arguments{
\item{model}{An irt_model object}

\item{nmtab_file}{path to the table file}
}
\value{
A data frame with item parameters
}
\description{
Read item paramter values from NM table
}
