% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{remove_items}
\alias{remove_items}
\title{Remove one or more items from a scale}
\usage{
remove_items(scale, numbers)
}
\arguments{
\item{scale}{An irt_scale object}

\item{numbers}{A vector of the item numbers to remove}
}
\value{
A new irt_scale object with the items removed
}
\description{
Remove one or more items from a scale
}
