% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_models.R
\name{renumber_parameters}
\alias{renumber_parameters}
\title{Renumber parameters used in an array of NM-TRAN abbreviated code lines}
\usage{
renumber_parameters(code, next_theta, next_eta)
}
\arguments{
\item{code}{An character vector of code lines}

\item{next_theta}{The number of the next THETA to be used}

\item{next_eta}{The number of the next ETA to be used}
}
\value{
A list of an array of new code lines, the next_theta and the next_eta
}
\description{
THETA(1) in the code will become next_theta, THETA(2) will become next_theta+1 etc
}
