% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_models.R
\name{reset_lv_models}
\alias{reset_lv_models}
\title{Reset all added latent variable models}
\usage{
reset_lv_models(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A new irt_model object
}
\description{
Reset all added latent variable models
}
