% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{response_probability_prediction_code}
\alias{response_probability_prediction_code}
\title{Generate NONMEM code for response probability and residual}
\usage{
response_probability_prediction_code()
}
\value{
A code generator object
}
\description{
Generate NONMEM code for response probability and residual
}
