% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{save_model}
\alias{save_model}
\title{Save NONMEM code}
\usage{
save_model(model, path)
}
\arguments{
\item{model}{A model object}

\item{path}{Path to the created model file}
}
\description{
Saves the NONMEM code of a model object to a file
}
