% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{save_scale}
\alias{save_scale}
\title{Save a scale to file}
\usage{
save_scale(scale, path)
}
\arguments{
\item{scale}{An irt_scale object}

\item{path}{Path to the file to save the scale}
}
\description{
\code{save_scale} saves a scale object to a file
}
