% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{scale_from_df}
\alias{scale_from_df}
\alias{scale_from_csv}
\title{Infer a scale from a dataset}
\usage{
scale_from_df(df, item = "ITEM", dv = "DV", name = NULL,
  type = NULL)

scale_from_csv(file, item = "ITEM", dv = "DV", name = NULL,
  type = NULL)
}
\arguments{
\item{df}{A data.frame}

\item{item}{Name of the item columns. Default is \code{ITEM}}

\item{dv}{Name of the dv column. Default is \code{DV}}

\item{name}{Name of an optional name column. Each item can be given a name (mainly for plotting) through this column. Only the first row of a new item will be used for the name.}

\item{type}{Name of an optional type column. Each item can be given a type (either ordcat or binary) through this column. If the type option isn't used the type is 
inferred from the number of distinct DV values}

\item{file}{The path to a csv file}
}
\value{
An irt_scale object
}
\description{
\code{scale_from_df} and \code{scale_from_csv} determine the scale from a dataset (provided as data.frame or csv file) 
using the ITEM, DV and MDV columns.
}
