% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{scale_overview}
\alias{scale_overview}
\title{Print an overview of a scale}
\usage{
scale_overview(scale)
}
\arguments{
\item{scale}{An irt_scale object}
}
\value{
A data.frame with scale information
}
\description{
Generate a table with the columns Item, Levels, Type, Categories and Name to give an
overview of what a scale object contains.
}
