% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{select_categories}
\alias{select_categories}
\title{Select categories}
\usage{
select_categories(scale, categories)
}
\arguments{
\item{scale}{An irt_scale object}

\item{categories}{A vector of category names to include in the new scale}
}
\value{
A new irt_scale object
}
\description{
\code{select_categories} Create a new scale with items from selected categories only
}
