% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{set_dataset}
\alias{set_dataset}
\title{Set dataset}
\usage{
set_dataset(
  model,
  path,
  use_path = TRUE,
  data_columns = NULL,
  mdv_column = NULL
)
}
\arguments{
\item{model}{A model object}

\item{path}{Path to a dataset file}

\item{use_path}{Should the path be put in $DATA or not? If FALSE only the filename will go into $DATA}

\item{data_columns}{Optional character vector of data columns for $INPUT to use instead of the dataset header}

\item{mdv_column}{Name of the column in the dataset to use a missing dependent variable indicator (MDV)}
}
\value{
A new model object
}
\description{
This function sets the dataset of a model (i.e., $DATA) as well as the data columns in $INPUT.
}
\details{
The function will open the dataset to extract the column names for $INPUT. The dataset is expected to be in CSV format without special characters
in the header.
}
