% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{set_initial_estimates}
\alias{set_initial_estimates}
\alias{set_initial_estimates_table}
\title{Set initial estimates}
\usage{
set_initial_estimates(model, items, parameters, inits)

set_initial_estimates_table(model, df)
}
\arguments{
\item{model}{An irt_model object}

\item{items}{A vector of item numbers}

\item{parameters}{A vector of parameter names that will be used for each of the items}

\item{inits}{A vector of initial estimates. Must be equal in size to parameters or have size one.}

\item{df}{A data.frame with columns 'item', 'parameter', 'init'}
}
\value{
A new irt_model object
}
\description{
These functions allow to set the initial estimates for parameters in the model. Values
can either be set for particular set of items and parameters or by providing a data.frame with 
the inital values for each parameter.
}
