% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{set_run_number}
\alias{set_run_number}
\title{Set the run number}
\usage{
set_run_number(model, run_number)
}
\arguments{
\item{model}{A model object}

\item{run_number}{An integer}
}
\description{
The run number will be added to the names of the table files
}
