% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{set_scale}
\alias{set_scale}
\title{Change the scale of an IRT model object}
\usage{
set_scale(model, scale)
}
\arguments{
\item{model}{An irt_model orbject}

\item{scale}{The scale object}
}
\value{
The updated model object
}
\description{
\code{set_scale} returns a newly created model object
}
