% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{simulation_code}
\alias{simulation_code}
\title{Generate NONMEM code for model simulation}
\usage{
simulation_code(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A code generator object
}
\description{
Generate NONMEM code for model simulation
}
