% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{simulation_task}
\alias{simulation_task}
\title{Generate NONMEM code for a model simulation task}
\usage{
simulation_task(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A code generator object
}
\description{
Generate NONMEM code for a model simulation task
}
