% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{str_irt_model}
\alias{str_irt_model}
\title{Create NONMEM model as a string}
\usage{
str_irt_model(model)
}
\arguments{
\item{model}{A model object}
}
\value{
A string with the NONMEM code
}
\description{
Create NONMEM model as a string
}
