% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{summary.irt_scale}
\alias{summary.irt_scale}
\title{Print a summary overview of a scale}
\usage{
\method{summary}{irt_scale}(object, ...)
}
\arguments{
\item{object}{An irt_scale object}

\item{...}{No additional arguments are supported}
}
\value{
A data.frame with scale information
}
\description{
Generate a table with the columns Item, Levels, Type, Categories and Name to give an
overview of what a scale object contains.
}
