% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{theta_init}
\alias{theta_init}
\title{Give a theta init string from limits}
\usage{
theta_init(init, lower, upper)
}
\description{
init=0 gives 0 FIX
 
 @param init The initial estimate
 @param lower The lower bound (not mandatory)
 @param upper The upper bound (not mandatory)
 @return A theta init string for NONMEM
 @keywords internal
}
