% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{theta_placeholder}
\alias{theta_placeholder}
\title{Generate NONMEM code for unused THETAS}
\usage{
theta_placeholder(cg, item)
}
\arguments{
\item{cg}{A code generator object to add lines to}

\item{item}{An irt item object to add placeholders for}
}
\value{
A new code generator object
}
\description{
Theta placeholders will be used for items that are in the base_scale
but not in the scale of a model. This is to keep theta numbering the same
for subscales and for full scales
}
