% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{theta_string_fix_part}
\alias{theta_string_fix_part}
\title{Generate the fix part of the $THETA record for an item}
\usage{
theta_string_fix_part(model, item, parameter)
}
\arguments{
\item{model}{An irt_model object}

\item{item}{An irt_item object}

\item{parameter}{The item parameter name}
}
\value{
A string containing "FIX" or an empty string
}
\description{
FIX should be returned if:
}
\details{
\itemize{
\item init = 0
\item The level corresponding to param was consolidated (in the upper end of the levels)
\item fix was set using the fix_item_parameters function
}
}
