% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{theta_string_init_part}
\alias{theta_string_init_part}
\title{Generate the init and limits part of the $THETA record for an item}
\usage{
theta_string_init_part(model, item, parameter)
}
\arguments{
\item{model}{An irt_model object}

\item{item}{An irt_item object}

\item{parameter}{The item parameter name}
}
\value{
A string containing the initial estimate and limits
}
\description{
Generate the init and limits part of the $THETA record for an item
}
