% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{theta_string_item_parameter}
\alias{theta_string_item_parameter}
\title{Create initial estimate $THETA for an item parameter}
\usage{
theta_string_item_parameter(model, item, parameter, theta_num)
}
\arguments{
\item{model}{An irt_model object}

\item{item}{An irt_item object}

\item{parameter}{The name of the parameter}

\item{theta_num}{The number of the THETA}
}
\description{
Create initial estimate $THETA for an item parameter
}
