% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{theta_string_label_part}
\alias{theta_string_label_part}
\title{Generate the label comment part of the $THETA record for an item}
\usage{
theta_string_label_part(model, item, parameter, theta_number)
}
\arguments{
\item{model}{An irt_model object}

\item{item}{An irt_item object}

\item{parameter}{The item parameter name}

\item{theta_number}{The number of the THETA}
}
\value{
A string containing the label for this THETA
}
\description{
Generate the label comment part of the $THETA record for an item
}
