% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{total_score_range}
\alias{total_score_range}
\title{Calculate the minimum and maximum total score for a scale}
\usage{
total_score_range(scale)
}
\arguments{
\item{scale}{An irt_scale object}
}
\value{
Vector of min and max total score
}
\description{
Calculate the minimum and maximum total score for a scale
}
