% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_model.R
\name{ts_model}
\alias{ts_model}
\title{Create a total score model object}
\usage{
ts_model(scale_or_min, max = NULL)
}
\arguments{
\item{scale_or_min}{A scale object or a minimum total score}

\item{max}{A maximum total score}
}
\value{
A model object
}
\description{
\code{ts_model} returns a newly created model object
}
