% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{type_constants}
\alias{type_constants}
\title{Generate the NONMEM code for definitions of type constants}
\usage{
type_constants(model)
}
\arguments{
\item{model}{An irt_model object}
}
\value{
A code generator object
}
\description{
Generate the NONMEM code for definitions of type constants
}
