% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{undef_item_code}
\alias{undef_item_code}
\title{Generate NONMEM code to clear variables for undef item}
\usage{
undef_item_code()
}
\value{
A code generator object
}
\description{
Generate NONMEM code to clear variables for undef item
}
