% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_or_insert}
\alias{update_or_insert}
\title{Update existing rows or insert new ones}
\usage{
update_or_insert(old, new, id_cols, only_existing_cols = TRUE)
}
\arguments{
\item{old}{the data.frame to update}

\item{new}{the data.frame with the updated values}

\item{id_cols}{columns used to join}

\item{only_existing_cols}{whether to allow new columns to be created}
}
\value{
data.frame with updated values
}
\description{
Update the rows in old to the ones in new if they exist, otherwise insert them as new rows. Rows are matched by the columns provided in id_cols.
}
