% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{update_parameters}
\alias{update_parameters}
\title{Update model parameters}
\usage{
update_parameters(model, data)
}
\arguments{
\item{model}{An irt_model object}

\item{data}{A data.frame}
}
\value{
An irt_model object
}
\description{
This function updates the parameter values of the model. It accepts data in long and wide
format and ,therefore, can be used to update from a NONMEM table or the output of 
\code{\link{estimate_item_parameters}}.
}
\details{
The long format is recognized by the presence of the column 'parameter'. In long format, only the columns
item, parameter and value are used.  

If the 'parameter' column is not found, the wide format is assumed for the parameter data. In wide format, the columns
item, dis, gue, dif, and difN (where N indicates a number) are used. All other columns are ignored. 

Capitalization of the column or parameter names does not matter.
}
