% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{add_dataset}
\alias{add_dataset}
\title{Add dataset}
\usage{
add_dataset(model, path)
}
\arguments{
\item{model}{A model object}

\item{path}{Path to a dataset file}
}
\value{
A new model object
}
\description{
Add a dataset to a model object
}
