% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_model_generics.R
\name{all_items}
\alias{all_items}
\alias{all_items.irt_scale}
\alias{all_items.irt_model}
\alias{items_by_type}
\alias{items_by_type.irt_scale}
\alias{items_by_type.irt_model}
\alias{items_in_categories}
\alias{items_in_categories.irt_scale}
\alias{items_in_categories.irt_model}
\title{Get a vector of item numbers}
\usage{
all_items(o)

\method{all_items}{irt_scale}(o)

\method{all_items}{irt_model}(o)

items_by_type(o, types)

\method{items_by_type}{irt_scale}(o, types)

\method{items_by_type}{irt_model}(o, types)

items_in_categories(o, categories)

\method{items_in_categories}{irt_scale}(o, categories)

\method{items_in_categories}{irt_model}(o, categories)
}
\arguments{
\item{o}{An irt_scale or irt_model object}

\item{types}{A vector of item types}

\item{categories}{A vector of names of categories}
}
\value{
A vector of item numbers
}
\description{
These function return a vector of item numbers, either for all items (\code{all_items}), all items with a certain type (\code{items_by_type}) 
or all items in a certain category (\code{items_in_categories}) in a scale or model.
}
