% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_model.R
\name{bi_model}
\alias{bi_model}
\title{Create a bounded integer model object}
\usage{
bi_model(scale_or_min, max = NULL)
}
\arguments{
\item{scale_or_min}{A scale object or a minimum total score}

\item{max}{A maximum total score}
}
\value{
A model object
}
\description{
\code{bi_model} returns a newly created model object
}
