% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{consolidate_levels}
\alias{consolidate_levels}
\title{Consolidate item levels}
\usage{
consolidate_levels(scale, item_number, levels)
}
\arguments{
\item{scale}{An irt_scale object}

\item{item_number}{The number of an item}

\item{levels}{The levels to be merged together}
}
\value{
A new irt_scale object
}
\description{
Consolidate, i.e. merge levels together, of a certain item in a scale object.
}
