% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{consolidate_levels}
\alias{consolidate_levels}
\title{Consolidate item levels}
\usage{
consolidate_levels(scale, item_numbers, levels)
}
\arguments{
\item{scale}{An irt_scale object}

\item{item_numbers}{A vector of item numbers to apply the same consolidation to}

\item{levels}{The levels to be merged together}
}
\value{
A new irt_scale object
}
\description{
Consolidate, i.e. merge levels together, of a certain items in a scale object. Consolidated items will be removed
from the scale.
}
