% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{correlation_plot}
\alias{correlation_plot}
\title{Item response correlation plot}
\usage{
correlation_plot(df)
}
\arguments{
\item{df}{A data.frame from the item_parameters_tab of a model run. ID, ITEM, TIME and PWRES will be used.}
}
\value{
A plot object
}
\description{
Visualizes the correlation of residuals between items as a correlation coefficients heatmap.
}
\details{
The residuals need to be available as PWRES in the data and are assumed to be standard Pearson residuals (i.e., difference between 
observed and expected devided by the expected standard devition). Models that are generated with this package will include the 
proper calculation of PWRES by default.
}
