% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{icc_plots}
\alias{icc_plots}
\title{Plot item characteristic curves}
\usage{
icc_plots(df, scale, items_per_page = 8)
}
\arguments{
\item{df}{A data.frame from the item_parameters_tab of a model run. ITEM, DV, DIS, DIFn are needed}

\item{scale}{The corresponding scale object}

\item{items_per_page}{Default to 8}
}
\value{
A list of pages
}
\description{
MDV and EVID will be used to filter out non-observations before plotting
}
\section{Warning}{

Binary items are not handled well.
}

