% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters.R
\name{initial_estimate}
\alias{initial_estimate}
\title{Get the initial estimate for a certain item and parameter}
\usage{
initial_estimate(model, item, parameter)
}
\arguments{
\item{model}{An irt_model object}

\item{item}{An irt_item object}

\item{parameter}{The name of a parameter}
}
\value{
The initial estimate of the parameter for the item
}
\description{
The initial estimate will be:
}
\details{
\enumerate{
\item The user defined value from initial_estimates_item_parameters
\item Taken from the published model in the scale
\item A fall back value of:
Binary: DIS=1, DIF=0.1, GUE=0.01
Ordered categorical: DIS=1, DIFn=-3 + 6n/#cats (not lower than 0.1)
}
}
