% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{item_inits}
\alias{item_inits}
\title{Create initial estimates string for item}
\usage{
item_inits(item, consolidated = NULL)
}
\arguments{
\item{item}{An item object}
}
\value{
A vector of initial estimates definitions for NONMEM
}
\description{
Create initial estimates string for item
}
\keyword{internal}
