% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{item_levels}
\alias{item_levels}
\title{Get vector of levels from character description or vector of levels}
\usage{
item_levels(x)
}
\arguments{
\item{x}{Can either be a string or an array of levels}
}
\value{
A sorted array of levels
}
\description{
If a vector of levels is input it will simply be output as it is
If a string is input it will be parsed to get a list of all levels it describes
Allowed formats are [a,b] for an interval and (1,2,3,4) to explicitly write all possible levels
}
\examples{
levels <- item_levels("[1,3]")
levels <- item_levels("(0,1,2,3,4)")
means 1, 2 and 3. 
}
\keyword{internal}
