% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{item_parameter_index}
\alias{item_parameter_index}
\title{Get the index of an item parameter relative to the first parameter of that item}
\usage{
item_parameter_index(scale, item_number, parameter)
}
\arguments{
\item{scale}{An irt_scale object}

\item{item_number}{Number of the item}

\item{parameter}{The name of the parameter}
}
\value{
The index of the parameter  starting from 1 or NA if not found
}
\description{
DIS would have index 1, DIF1 would have 2 etc
}
