% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{item_parameter_names}
\alias{item_parameter_names}
\title{Get vector of parameter names of an item}
\usage{
item_parameter_names(scale, item_number)
}
\arguments{
\item{scale}{An irt_scale object}

\item{item_number}{An item number}
}
\value{
A character vector of parameter names
}
\description{
Get vector of parameter names of an item
}
