% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mirror_plots}
\alias{mirror_plots}
\title{Mirror plots for comparison of original data and simulated data}
\usage{
mirror_plots(origdata, model, simdata = NULL, nrow = 4, ncol = 5)
}
\arguments{
\item{origdata}{The original dataset}

\item{model}{The model used to generate the data}

\item{simdata}{The simulated data. Will plot only original data if this is missing}

\item{nrow}{The number of rows per page to use for the matrix of plots}

\item{ncol}{The number of columns per page to use for the matrix of plots}
}
\value{
A list of plots. One page per item.
}
\description{
MDV and EVID will be used to filter out non-observations before plotting
}
