% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmIRT.R
\docType{package}
\name{nmIRT-package}
\alias{nmIRT}
\alias{nmIRT-package}
\title{nmIRT: a package for creating and diagnosing IRT models for NONMEM}
\description{
Automatic generation of different types of item response theory (IRT) models.
    Diagnostic tools for IRT modelling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UUPharmacometrics/nmIRT}
  \item Report bugs at \url{https://github.com/UUPharmacometrics/nmIRT/issues}
}

}
\author{
\strong{Maintainer}: Rikard Nordgren \email{rikard.nordgren@farmbio.uu.se} [copyright holder]

Authors:
\itemize{
  \item Leticia Arrington [copyright holder]
  \item Sebastian Ueckert [copyright holder]
  \item Mats O. Karlsson [copyright holder]
}

}
