% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{scale_from_dataset}
\alias{scale_from_dataset}
\title{Infer a scale from a dataset}
\usage{
scale_from_dataset(df, item = "ITEM", dv = "DV")
}
\arguments{
\item{df}{Can either be a data.frame or a path to a csv file}

\item{item}{Name of the item columns. Default is \code{ITEM}}

\item{dv}{Name of the dv column. Default is \code{DV}}
}
\value{
An irt_scale object
}
\description{
\code{scale_from_dataset} determines the scale from a dataset using the ITEM, DV and MDV columns.
All items will be assumed to be of type ordered categorical
}
