% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{set_dataset}
\alias{set_dataset}
\title{Set dataset}
\usage{
set_dataset(model, path, use_path = TRUE)
}
\arguments{
\item{model}{A model object}

\item{path}{Path to a dataset file}

\item{use_path}{Should the path be put in $DATA or not? If FALSE only the filename will go into $DATA}
}
\value{
A new model object
}
\description{
Set the dataset of a model object
}
