% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model.R
\name{set_scale}
\alias{set_scale}
\title{Change the scale and/or base scale of an IRT model object}
\usage{
set_scale(model, scale, base_scale)
}
\arguments{
\item{scale}{The scale}

\item{base_scale}{A base scale on which the scale is based}
}
\value{
The updated model object
}
\description{
\code{set_scale} returns a newly created model object
}
