% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_scale.R
\name{theta_init}
\alias{theta_init}
\title{Give a theta init string from limits}
\usage{
theta_init(init, lower, upper)
}
\arguments{
\item{init}{The initial estimate}

\item{lower}{The lower bound (not mandatory)}

\item{upper}{The upper bound (not mandatory)}
}
\value{
A theta init string for NONMEM
}
\description{
init=0 gives 0 FIX
}
\keyword{internal}
